fix(material/form-field): Correct label position problem with outline appearance inside material tabs
This change helps with an issue when you have a form field with an outline appearance placed inside a material tab. When the form field has an outline, the label does not show up properly. By adding some space at the top when the form field has an outline, we can make sure the label is not cut off.

Fixes #27764