fix(material/autocomplete): autocomplete panel top is cut off in landscape mode
Fixes lint error from previous commit which fixes Angular Component's
autocomplete panel top from being cut off when viewed in landscape mode.
The previous max-height of the panel is more than the height of the panel
from the top of the input when in the bottom half to the top of the device's
screen.

Fixes b/284148377